Modifying spreadsheets now works.
Created google.js, responsible for logging into the service account.
Created SpreadsheetDatabase class, for interacting with a sheets
document in an array-like manner. You can get / set slices or individual
rows of data, get / set individual parts of a row, and get all values of
a particuar column.
service-account-data.json has all the necessary data to log in to the
service account, and should thus not be made available outside of the
server.