Google sign-in is better, steps look nicer.
Gave individual steps their own cards that hide most of their content
until they are needed. Also checkboxes.
As a byproduct, material icons are now available.
Added a kind of hacky solution for people who have never signed in to
their google account on the browser they are using, but since this is so
rare, it is not a priority to smooth out that experience.