Made so many things better and nicer all around.
The steps are now numbered.
Checkboxes only show up when steps are completed, so people don't try to
click on them.
Polished the manual sign in experience a lot, since it is also required
if you have an alternate account that you want to use but is not listed.
(Added a link to do that as well, even when you already have an account
listed in the YOLO box.)
Added a proper placeholder message to the homepage.
Fixed that weird scrolling bug, everything scrolls the right amount now.
Added a list of useful information that shows up when someone is done
signing in.
Added a check to see if a user already registered as soon as they sign
in, rather than after they fill out the form.
To make this work, added an endpoint to retreive user information from
their email address. (Returns 404 if no user has the specified email
address.)
Removed the 'You already registered' message and just made it check the
second step right away instead.
Added special handling for boolean values in sheetDatabase.
Refactored the backend service to use promises instead of yucky
callbacks.