Reorganized stuff.
Replaced home button with new logo.
Reorganized pages into 'public' and 'private' folders.
Made the slideshow and centering div part of the app component.
Made the slideshow show up on all public pages.
Added a variable to WebappBackendService to keep track of whether or not
the user is signed in and registered.
Added different sets of buttons for the title bar depending on whether
or not the user is signed in and registered.