2/?, aesthetic changes
Not much here:
Added a less placeholdery but still placeholdery home page with a
stylized 'FRC 5669' logo and a link to the FRC website.
Moved the centered layout from being applied to all pages to being a
simple component that can be wrapped around whichever pages need it.
This will allow for using alternate layouts for denser pages when they
are made.
Added a very placeholdery members page which will eventually be used to
show people in the team.