3/3, misc changes
Added server-side stuff to google.js for easy validation of YOLO tokens.
Modified the recruit page logic to be compliant with the new API
endpoints.
Modified logic of the top bar to be compliant with the modified wrapper
incorporating the new API endpoints.
Modified WebappBackendService code to be much cleaner.
Added automatic authorization header insertion in WebappBackendService.