Implemented proper member viewing / editing
Added a component for viewing / editing details of a member.
It shows information when the user is a member and allows editing when
the user is an admin or the user is editing themself.
It can be accessed through the path /private/members/:email
Made the members grid show a bit more information per user.
Made the members grid show whether the member can only view or can also
edit each member.
Added spacing to the 'send me emails' checkbox on the sign up page.
Made the site try to automatically log you in when it is first loaded.