Implemented server-side caching of spreadsheet files.
It was painful.

All methods in SpreadsheetDatabase now work off of a local copy of the
database, making them much faster.
All reading methods in SpreadsheetDatabase will check if it has been
longer than 5 seconds since a local copy of the spreadsheet was
downloaded. If so, a new copy will be downloaded.
All writing methods will schedule an upload of the changes to happen 2
seconds after the method completes. If another write method is used
inside that time span, the upload is pushed back to 2 seconds after that
latest method finishes.
Fixed a bug somewhere probably.