Security
HTTPS is finally working!
Server now tells everyone connecting via http on port 80 to go to https
on port 443
Server can now be switched between production and development mode, by
passing 'prod' as an argument to node.
Without the 'prod' argument, website will be served via http on port
4200.
Deleted service-account-data.json from the repo, because it has an
important private key in it. Created a new one which is now secret and
no longer visible to anyone who can see the repo.