Reworked how the login code works on the frontend.
Yolo service now provides access to login details via non-blocking
synchronous and asynchronous methods.
Yolo service now attempts to log in the user automatically as soon as
the page is loaded (which was previously part of the webapp backend
service.)
Removed a bunch of different methods for logging in using Yolo service,
to simplify things.
The only remaining login method, hint(), works by calling Yolo's hint as
soon as the previous login attempt (either automatic or another hint)
has failed.
hint() returns a promise that is resolved / rejected when the user logs
in / the hint fails.

WebappBackendService now works similarly to the changes in YoloService:
WBS now has methods for accessing current member's data synchronously
without blocking and asynchronously.
WBS now attaches to the asynchronous retrieval of google credentials in
the Yolo service to log in users instead of relying on starting
automatic login itself or on special code in the recruit page.

Refactored a bunch of stuff to use this new, more streamlined system.