User experience improvements
Fixed spacing problems in app-grid-layout, it now looks more uniform.
Added the nifty 'app-page-guard' component which handles displaying
messages for pages where content needs to load and / or requires access.
Added app-page-guard to members and member pages.
Changed 'admin' to 'leader' to make it more obvious who gets that rank.
Added a 'NYI' message on the Parts Request page.
Now when a user registers, a link to their profile picture is saved and
is displayed on the members and member pages.
Compressed the background slides shown on public pages to be more than 3
times smaller (in file size) than before to load faster over slow
connections.