Few small cosmetic changes
The info that is displayed when a user first registers can now be
accessed when they are logged in by clicking the new 'info' button in
the title bar.
Added favicons that take advantage of the various perks different
platforms provide (macos touchbar, windows start menu, chrome on android
shortcuts, etc.)
Moved icons of Google Drive and Team Reach to be hosted by us instead of
an outside source.
Added a hint on the info page to point the user to the members page to
see who everyone is.