Added 'sessionStorage'.
It stores sessions of users using the website with associated tokens.
These sessions can store the ID of a user that the session user has
proved they own the account of. The client can thus use the session
token to prove their identity. Currently, they are deleted upon
restarting the server. Will need to add some sort of expiration for them
in the future.